Refactor `src/packages/` folder into file
